fix Big Sum Submissions

+++ problems/big_sum/submissions/time_limit_exceeded/big_sum_time_limit_exceeded.c
@@ -2,12 +2,12 @@
 
 int main(int argc, char const *argv[])
 {
-    int n;
-    scanf("%d", &n);
-    int s = 0;
-    for (int i = 0; i <= n; i++){
+    long n;
+    scanf("%ld", &n);
+    long s = 0;
+    for (long i = 0; i <= n; i++){
         s += i;
     }
-    printf("%d", s);
+    printf("%ld", s);
     return 0;
 }
